requirements.txt
appdirs==1.4.0
click==6.7
Flask==0.12
Flask-MySQL==1.4.0
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
packaging==16.8
pycrypto==2.6.1
PyMySQL==0.7.9
pyparsing==2.1.10
requests==2.13.0
selenium==3.0.2
six==1.10.0
Werkzeug==0.11.15
